Fix import of datematch in test.py
Has moved to utils.py from bake.py. Also rename the test class since,
(un)hilariously, it hid the imported name.
